+++ app.js
@@ -831,6 +831,8 @@ import { renderers as RV1476 } from './views/group-v1476.js';
 import { renderers as RV1479 } from './views/group-v1479.js';
 import { renderers as RV1480 } from './views/group-v1480.js';
 import { renderers as RV1481 } from './views/group-v1481.js';
+import { renderers as RV1482 } from './views/group-v1482.js';
+import { renderers as RV1483 } from './views/group-v1483.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -919,7 +921,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV1479, ...RV1480, ...RV1481, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV1479, ...RV1480, ...RV1481, ...RV1482, ...RV1483, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -2206,7 +2208,9 @@ const UTILITIES = [
   { id: 'bpe-periodontal', name: 'Basic Periodontal Examination (BPE)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'loe-silness-gingival-index', name: 'Loe-Silness Gingival Index', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'silness-loe-plaque-index', name: 'Silness-Loe Plaque Index', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'ohi-s', name: 'Simplified Oral Hygiene Index (OHI-S)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'miller-gingival-recession', name: 'Miller Gingival Recession Class', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'cairo-recession', name: 'Cairo Classification (Gingival Recession)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'glickman-furcation', name: 'Glickman Furcation Grade', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'isi', name: 'Insomnia Severity Index (ISI)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'fois', name: 'Functional Oral Intake Scale (FOIS)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },